move headbar to modules
